requirements.txt
fastapi
uvicorn[standard]
langchain>=0.1.15
openai>=1.22.0
python-dotenv>=1.0.1
langchain_openai
langchain_community
langchain-google-genai
langchain-groq
streamlit